Move common tests to a new file
